validaciones y pruebas Modulo Usuario

+++ src/main/java/com/api/modules/user/service/UserService.java
@@ -5,11 +5,11 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.api.modules.user.dto.UserRequestDTO;
 import com.api.modules.user.dto.UserResponseDTO;
-import com.api.modules.user.factory.UserFactory;
 import com.api.modules.user.mapper.UserMapper;
 import com.api.modules.user.model.User;
 import com.api.modules.user.repository.UserRepository;
@@ -34,22 +34,20 @@ public class UserService {
     public UserResponseDTO getById(UUID id) {
         return userRepository.findById(id)
                 .map(UserMapper::toResponseDTO)
-                .orElse(null);
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
     }
 
     // Crear un nuevo usuario (mas adelante cifrado de contraseña)
     public UserResponseDTO create(UserRequestDTO dto) {
-        // Convertimos el DTO a entidad :,D
+        if (userRepository.existsByEmail(dto.getEmail())) {
+            throw new DataIntegrityViolationException("El correo electrónico ya está registrado.");
+        }
+
         User user = UserMapper.toEntity(dto);
-        user = UserFactory.createUser(
-                user.getName(),
-                user.getEmail(),
-                user.getPassword());
-        // Guardamos en la BD
         User saved = userRepository.save(user);
-        // Retornamos la respuesta como DTO
         return UserMapper.toResponseDTO(saved);
     }
+
 
     // Actualizar usuario existente
     public UserResponseDTO update(UUID id, UserRequestDTO dto) {
@@ -62,11 +60,14 @@ public class UserService {
                     User updated = userRepository.save(user);
                     return UserMapper.toResponseDTO(updated);
                 })
-                .orElse(null);
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado para actualizar"));
     }
 
     // Eliminar usuario por ID
     public void delete(UUID id) {
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("Usuario no encontrado para eliminar");
+        }
         userRepository.deleteById(id);
     }
 }
